Switched to using long double.

+++ include/core.h
@@ -3,15 +3,15 @@
 
 #include <stdbool.h>
 
-double decimal_reciprocal(int n, int pos);
-double compute_3_4_first_sum(int n, int base, int c, int (*p)(int), bool start_at_0);
-double compute_3_4_second_sum(int n, int base, int c, int (*p)(int));
-double mod_one(double x);
-char* convert_floating_decimal_to_hex(double x, int places);
-char* convert_floating_decimal_to_base(double x, int places, int base);
-double compute_3_4(int n, int base, int c, int (*p)(int), bool start_at_0);
+long double decimal_reciprocal(int n, int pos);
+long double compute_3_4_first_sum(int n, int base, int c, int (*p)(int), bool start_at_0);
+long double compute_3_4_second_sum(int n, int base, int c, int (*p)(int));
+long double mod_one(long double x);
+char* convert_floating_decimal_to_hex(long double x, int places);
+char* convert_floating_decimal_to_base(long double x, int places, int base);
+long double compute_3_4(int n, int base, int c, int (*p)(int), bool start_at_0);
 char* compute_3_4_to_base(int n, int base, int c, int (*p)(int), int places, bool start_at_0);
-double calculate_machine_epsilon(void);
+long double calculate_machine_epsilon(void);
 
 #endif // CORE_H
