+++ include/transcendental.h
@@ -8,10 +8,10 @@ int pi_hex_p1(int k);
 int pi_hex_p2(int k);
 int pi_hex_p3(int k);
 int pi_hex_p4(int k);
-double compute_pi_sum1(int first_place, int num_places);
-double compute_pi_sum2(int first_place, int num_places);
-double compute_pi_sum3(int first_place, int num_places);
-double compute_pi_sum4(int first_place, int num_places);
+long double compute_pi_sum1(int first_place, int num_places);
+long double compute_pi_sum2(int first_place, int num_places);
+long double compute_pi_sum3(int first_place, int num_places);
+long double compute_pi_sum4(int first_place, int num_places);
 
 #endif // TRANSCENDENTAL_H
